move deep link to code editor only
